Prefill cache of prix carburants.
