+++ openfisca_france_indirect_taxation/variables/taxes_indirectes/prix_carburants_ttc_AN.py
@@ -1,5 +1,5 @@
 from openfisca_france_indirect_taxation.variables.base import Menage, Variable, YEAR
-from openfisca_france_indirect_taxation.variables.prix_carburants_regions_years_AN import get_prix_carburant_par_region_par_carburant_par_an_hectolitre
+from openfisca_france_indirect_taxation.variables.taxes_indirectes.prix_carburants_par_region import get_prix_carburant_par_region_par_carburant_par_an_hectolitre
 
 import numpy as np
